Address code review feedback.
This change incorporates code review feedback, including:

- Refactoring BigQueryIO.TypedRead.expand to break up the method
- Modifying temporary dataset cleanup to use PassThroughThenCleanup
  rather than Count.globally() to gate the cleanup process
- Removing an unused PTransform in the BigQueryWordCount sample
- Moving helper code to execute queries and create read sessions
  out of the related Source types and into helper classes
- Deleting the now-unused BigQueryParallelReadQuerySource and
  BigQueryParallelReadSourceBase classes
- Refactoring BigQueryParallelReadStreamReader to be a nested
  subclass of BigQueryParallelReadStreamSource, as is common with
  other Source types.
