+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryHelpers.java
@@ -25,8 +25,12 @@ import com.google.api.services.bigquery.model.JobStatus;
 import com.google.api.services.bigquery.model.TableReference;
 import com.google.api.services.bigquery.model.TableSchema;
 import com.google.api.services.bigquery.model.TimePartitioning;
+import com.google.cloud.bigquery.v3.ParallelRead;
+import com.google.cloud.bigquery.v3.ReadOptions;
+import com.google.cloud.bigquery.v3.TableReferenceProto;
 import com.google.cloud.hadoop.util.ApiErrorExtractor;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Strings;
 import com.google.common.hash.Hashing;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -37,7 +41,9 @@ import java.util.regex.Matcher;
 import javax.annotation.Nullable;
 import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.io.fs.ResolveOptions;
+import org.apache.beam.sdk.io.gcp.bigquery.BigQueryIO.ReadSessionOptions;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryServices.DatasetService;
+import org.apache.beam.sdk.io.gcp.bigquery.BigQueryServices.TableReadService;
 import org.apache.beam.sdk.options.ValueProvider;
 import org.apache.beam.sdk.options.ValueProvider.NestedValueProvider;
 import org.apache.beam.sdk.transforms.SerializableFunction;
@@ -242,6 +248,47 @@ public class BigQueryHelpers {
       jobId += String.format("_%05d", index);
     }
     return jobId;
+  }
+
+  static ParallelRead.Session createReadSession(
+      TableReadService tableReadService,
+      TableReference tableReference,
+      int readerCount,
+      ReadSessionOptions readSessionOptions) {
+
+    ParallelRead.CreateSessionRequest.Builder requestBuilder =
+        ParallelRead.CreateSessionRequest.newBuilder()
+            .setTableReference(
+                TableReferenceProto.TableReference.newBuilder()
+                    .setProjectId(tableReference.getProjectId())
+                    .setDatasetId(tableReference.getDatasetId())
+                    .setTableId(tableReference.getTableId()))
+            .setReaderCount(readerCount);
+
+    if (readSessionOptions != null) {
+      ReadOptions.TableReadOptions.Builder readOptionsBuilder = null;
+      String sqlFilter = readSessionOptions.getSqlFilter();
+      if (!Strings.isNullOrEmpty(sqlFilter)) {
+        readOptionsBuilder = ReadOptions.TableReadOptions.newBuilder()
+            .setSqlFilter(sqlFilter);
+      }
+
+      List<String> selectedFields = readSessionOptions.getSelectedFields();
+      if (selectedFields != null && !selectedFields.isEmpty()) {
+        if (readOptionsBuilder == null) {
+          readOptionsBuilder = ReadOptions.TableReadOptions.newBuilder();
+        }
+        for (String selectedField : selectedFields) {
+          readOptionsBuilder.addSelectedFields(selectedField);
+        }
+      }
+
+      if (readOptionsBuilder != null) {
+        requestBuilder.setReadOptions(readOptionsBuilder);
+      }
+    }
+
+    return tableReadService.createSession(requestBuilder.build());
   }
 
   @VisibleForTesting
