+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryParallelReadTableSource.java
@@ -23,8 +23,12 @@ import static com.google.common.base.Preconditions.checkState;
 
 import com.google.api.services.bigquery.model.Table;
 import com.google.api.services.bigquery.model.TableReference;
+import com.google.cloud.bigquery.v3.ParallelRead;
 import com.google.common.base.Strings;
+import com.google.common.collect.ImmutableList;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.io.BoundedSource;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryHelpers.TableRefToJson;
@@ -39,9 +43,23 @@ import org.slf4j.LoggerFactory;
 /**
  * A {@link BoundedSource} for reading BigQuery tables using the BigQuery parallel read API.
  */
-class BigQueryParallelReadTableSource<T> extends BigQueryParallelReadSourceBase<T> {
+class BigQueryParallelReadTableSource<T> extends BoundedSource<T> {
 
   private static final Logger LOG = LoggerFactory.getLogger(BigQueryParallelReadTableSource.class);
+
+  /**
+   * The maximum number of readers which will be requested when creating a BigQuery read session as
+   * part of a split operation, regardless of the desired bundle size.
+   */
+  private static final int MAX_SPLIT_COUNT = 10000;
+
+  /**
+   * The minimum number of readers which will be requested when creating a BigQuery read session as
+   * part of a split operation, regardless of the desired bundle size. Note that the source may
+   * still be split into fewer than ten component sources depending on the number of read locations
+   * returned by the server at session creation time.
+   */
+  private static final int MIN_SPLIT_COUNT = 10;
 
   /**
    * This method creates a new {@link BigQueryParallelReadTableSource} with no initial read session
@@ -64,6 +82,10 @@ class BigQueryParallelReadTableSource<T> extends BigQueryParallelReadSourceBase<
   }
 
   private final ValueProvider<String> jsonTableRefProvider;
+  private final BigQueryServices bqServices;
+  private final SerializableFunction<SchemaAndRowProto, T> parseFn;
+  private final Coder<T> coder;
+  private final BigQueryIO.ReadSessionOptions readSessionOptions;
 
   private transient Long cachedReadSizeBytes;
 
@@ -73,27 +95,55 @@ class BigQueryParallelReadTableSource<T> extends BigQueryParallelReadSourceBase<
       Coder<T> coder,
       BigQueryServices bqServices,
       ReadSessionOptions readSessionOptions) {
-    super(bqServices, parseFn, coder, readSessionOptions);
+    this.bqServices = checkNotNull(bqServices, "bqServices");
+    this.parseFn = checkNotNull(parseFn, "parseFn");
+    this.coder = checkNotNull(coder, "coder");
+    this.readSessionOptions = readSessionOptions;
     this.jsonTableRefProvider = checkNotNull(jsonTableRefProvider, "jsonTableRefProvider");
   }
 
-  /**
-   * Returns the effective table reference for the table. If the caller has not specified a project
-   * ID directly in the table reference, then the project ID is set from the pipeline options.
-   */
-  protected TableReference getEffectiveTableReference(BigQueryOptions options) throws IOException {
-    TableReference tableReference =
-        BigQueryIO.JSON_FACTORY.fromString(jsonTableRefProvider.get(), TableReference.class);
-    if (Strings.isNullOrEmpty(tableReference.getProjectId())) {
-      checkState(
-          !Strings.isNullOrEmpty(options.getProject()),
-          "No project ID was set in %s or %s; cannot construct a complete %s",
-          TableReference.class.getSimpleName(),
-          BigQueryOptions.class.getSimpleName(),
-          TableReference.class.getSimpleName());
-      tableReference.setProjectId(options.getProject());
+  @Override
+  public List<BoundedSource<T>> split(long desiredBundleSizeBytes, PipelineOptions options)
+      throws Exception {
+    BigQueryOptions bqOptions = options.as(BigQueryOptions.class);
+    TableReference tableReference = getEffectiveTableReference(bqOptions);
+
+    long tableSizeBytes = getEstimatedSizeBytes(options);
+    int readerCount = 0;
+    if (desiredBundleSizeBytes > 0) {
+      readerCount = (tableSizeBytes / desiredBundleSizeBytes) > MAX_SPLIT_COUNT
+          ? MAX_SPLIT_COUNT
+          : (int) (tableSizeBytes / desiredBundleSizeBytes);
     }
-    return tableReference;
+
+    if (readerCount < MIN_SPLIT_COUNT) {
+      readerCount = MIN_SPLIT_COUNT;
+    }
+
+    ParallelRead.Session readSession = BigQueryHelpers.createReadSession(
+        bqServices.getTableReadService(bqOptions),
+        tableReference,
+        readerCount,
+        readSessionOptions);
+
+    if (readSession.getInitialReadLocationsCount() == 0) {
+      return ImmutableList.of();
+    }
+
+    Long readSizeBytes = tableSizeBytes / readSession.getInitialReadLocationsCount();
+    List<BoundedSource<T>> sources = new ArrayList<>(readSession.getInitialReadLocationsCount());
+    for (ParallelRead.ReadLocation readLocation : readSession.getInitialReadLocationsList()) {
+      sources.add(new BigQueryParallelReadStreamSource<>(
+          bqServices,
+          parseFn,
+          coder,
+          readSessionOptions,
+          readSession,
+          readLocation,
+          readSizeBytes));
+    }
+
+    return ImmutableList.copyOf(sources);
   }
 
   /**
@@ -116,5 +166,34 @@ class BigQueryParallelReadTableSource<T> extends BigQueryParallelReadSourceBase<
     }
 
     return cachedReadSizeBytes;
+  }
+
+  @Override
+  public BoundedReader<T> createReader(PipelineOptions options) {
+    throw new UnsupportedOperationException("BigQuery source must be split before being read");
+  }
+
+  @Override
+  public Coder<T> getOutputCoder() {
+    return coder;
+  }
+
+  /**
+   * Returns the effective table reference for the table. If the caller has not specified a project
+   * ID directly in the table reference, then the project ID is set from the pipeline options.
+   */
+  private TableReference getEffectiveTableReference(BigQueryOptions options) throws IOException {
+    TableReference tableReference =
+        BigQueryIO.JSON_FACTORY.fromString(jsonTableRefProvider.get(), TableReference.class);
+    if (Strings.isNullOrEmpty(tableReference.getProjectId())) {
+      checkState(
+          !Strings.isNullOrEmpty(options.getProject()),
+          "No project ID was set in %s or %s; cannot construct a complete %s",
+          TableReference.class.getSimpleName(),
+          BigQueryOptions.class.getSimpleName(),
+          TableReference.class.getSimpleName());
+      tableReference.setProjectId(options.getProject());
+    }
+    return tableReference;
   }
 }
