+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryIOParallelReadTest.java
@@ -386,50 +386,6 @@ public class BigQueryIOParallelReadTest {
     assertEquals(100, source.getEstimatedSizeBytes(options));
   }
 
-  @Test
-  public void testQuerySourceEstimatedSize() throws Exception {
-    String stepUuid = "testStepUuid";
-
-    List<TableRow> expected = ImmutableList.of(
-        new TableRow().set("name", "a").set("number", 1L),
-        new TableRow().set("name", "b").set("number", 2L),
-        new TableRow().set("name", "c").set("number", 3L),
-        new TableRow().set("name", "d").set("number", 4L),
-        new TableRow().set("name", "e").set("number", 5L),
-        new TableRow().set("name", "f").set("number", 6L));
-
-    String encodedQuery = FakeBigQueryServices.encodeQuery(expected);
-
-    fakeJobService.expectDryRunQuery(
-        DEFAULT_PROJECT_ID,
-        encodedQuery,
-        new JobStatistics().setQuery(new JobStatistics2().setTotalBytesProcessed(100L)));
-
-    BoundedSource<Row> source = BigQueryParallelReadQuerySource.create(
-        fakeBigQueryServices,
-        SchemaAndRowProto::getRow,
-        ProtoCoder.of(Row.class),
-        null,
-        stepUuid,
-        ValueProvider.StaticValueProvider.of(encodedQuery),
-        false,
-        false);
-
-    // Get the estimated size once and verify that the bytes processed value from the dry run
-    // query is returned.
-    BigQueryOptions options = PipelineOptionsFactory.fromArgs("--project=" + DEFAULT_PROJECT_ID)
-        .as(BigQueryOptions.class);
-    assertEquals(100L, source.getEstimatedSizeBytes(options));
-
-    // Now verify that the cached value is returned.
-    fakeJobService.expectDryRunQuery(
-        DEFAULT_PROJECT_ID,
-        encodedQuery,
-        new JobStatistics().setQuery(new JobStatistics2().setTotalBytesProcessed(200L)));
-
-    assertEquals(100L, source.getEstimatedSizeBytes(options));
-  }
-
   private static final long TABLE_SIZE_BYTES = 1024L * 1024L;
 
   @Test
@@ -543,102 +499,6 @@ public class BigQueryIOParallelReadTest {
 
     List<? extends BoundedSource<Row>> sources = source.split(1024, options);
     assertEquals(0, sources.size());
-  }
-
-  @Test
-  public void testQuerySourceInitialSplit() throws Exception {
-    runTestQuerySourceInitialSplit(1024, 1024);
-  }
-
-  @Test
-  public void testQuerySourceInitialSplitMaxReaderCount() throws Exception {
-    runTestQuerySourceInitialSplit(16L, 10000);
-  }
-
-  @Test
-  public void testQuerySourceInitialSplitMinReaderCount() throws Exception {
-    runTestQuerySourceInitialSplit(TABLE_SIZE_BYTES, 10);
-  }
-
-  @Test
-  public void testQuerySourceInitialSplitBundleSizeZero() throws Exception {
-    runTestQuerySourceInitialSplit(0L, 10);
-  }
-
-  private void runTestQuerySourceInitialSplit(
-      long desiredBundleSizeBytes, int expectedReaderCount) throws Exception {
-
-    BigQueryOptions bqOptions =
-        PipelineOptionsFactory.fromArgs("--project=" + DEFAULT_PROJECT_ID)
-            .as(BigQueryOptions.class);
-
-    String stepUuid = "testStepUuid";
-
-    TableReference tempTableReference = createTempTableReference(
-        DEFAULT_PROJECT_ID, createJobIdToken(bqOptions.getJobName(), stepUuid));
-
-    fakeDatasetService.createDataset(
-        tempTableReference.getProjectId(),
-        tempTableReference.getDatasetId(),
-        "location",
-        "description",
-        null);
-
-    fakeDatasetService.createTable(new Table()
-        .setTableReference(tempTableReference)
-        .setSchema(new TableSchema()
-            .setFields(
-                ImmutableList.of(
-                    new TableFieldSchema().setName("name").setType("STRING"),
-                    new TableFieldSchema().setName("number").setType("INTEGER")))));
-
-    List<TableRow> expected = ImmutableList.of(
-        new TableRow().set("name", "a").set("number", 1L),
-        new TableRow().set("name", "b").set("number", 2L),
-        new TableRow().set("name", "c").set("number", 3L),
-        new TableRow().set("name", "d").set("number", 4L),
-        new TableRow().set("name", "e").set("number", 5L),
-        new TableRow().set("name", "f").set("number", 6L));
-
-    String encodedQuery = FakeBigQueryServices.encodeQuery(expected);
-
-    fakeJobService.expectDryRunQuery(
-        DEFAULT_PROJECT_ID,
-        encodedQuery,
-        new JobStatistics().setQuery(
-            new JobStatistics2()
-                .setTotalBytesProcessed(TABLE_SIZE_BYTES)
-                .setReferencedTables(ImmutableList.of(tempTableReference))));
-
-    CreateSessionRequest createSessionRequest = CreateSessionRequest.newBuilder()
-        .setTableReference(TableReferenceProto.TableReference.newBuilder()
-            .setProjectId(tempTableReference.getProjectId())
-            .setDatasetId(tempTableReference.getDatasetId())
-            .setTableId(tempTableReference.getTableId()))
-        .setReaderCount(expectedReaderCount)
-        .build();
-
-    Session.Builder sessionBuilder = Session.newBuilder().setName("session");
-    for (int i = 0; i < 50; i++) {
-      sessionBuilder.addInitialReadLocations(ReadLocation.newBuilder());
-    }
-
-    fakeTableReadService.setCreateSessionResult(createSessionRequest, sessionBuilder.build());
-
-    BoundedSource<Row> source = BigQueryParallelReadQuerySource.create(
-        fakeBigQueryServices,
-        SchemaAndRowProto::getRow,
-        ProtoCoder.of(Row.class),
-        null,
-        stepUuid,
-        ValueProvider.StaticValueProvider.of(encodedQuery),
-        false,
-        false);
-
-    List<? extends BoundedSource<Row>> sources = source.split(desiredBundleSizeBytes, bqOptions);
-    assertEquals(50, sources.size());
-    long expectedSizeBytes = TABLE_SIZE_BYTES / 50;
-    assertEquals(expectedSizeBytes, sources.get(0).getEstimatedSizeBytes(bqOptions));
   }
 
   @Test
